+++ pkg/vector_support_vectors.go
@@ -1,0 +1,5 @@
+//go:build vectors
+
+package pkg
+
+const vectorSupportEnabled = true